Implement prepend_first_pass option
In totpcgi deployments it's sometimes handy to ask for password and
token code separately, but pass them together to the totpcgi server --
e.g. if your token secrets are encrypted with users' pincodes. Adding
"prepend_first_pass" to the pam line will take the previously used
password in the stack and prepend it to the token code provided by the
user.